Moving do_help() to new do_utils.py file inside a new class called Do_Fuc()
