+++ easyaccess/eautils/__init__.py
@@ -1,2 +1,2 @@
 __all__ = ["db_api", "des_logo", "dircache", "dtypes",
-           "fileio", "fun_utils", "import_utils", "ea_utils"]
+           "fileio", "fun_utils", "import_utils", "ea_utils","do_utils"]
